updated readme, moved uploadTest to test folder, moved s3 methods to S3 folder, added gong favicon
