+++ README.md
@@ -1,13 +1,23 @@
 # niyagapedia
 An encyclopedia of Javanese gamelan musicians
 
-Todo:
+# Background
+The Solonese gamelan music scene has suffered the loss of many famous musicians in the past few years. I hope this simple record can help document notable musicians to keep their memories alive permanently.
 
+# A note about the app
+This project was also a vehicle for me to teach myself Express, using Pug as the view engine, and was the first project I created with either of these technologies. A lot of the code is therefore cobbled together from various tutorials, and in places I've used it as a sandbox for playing around and having fun.
 
-make sure all basic features are working
+# In progress
+Migration from Cloudinary to S3 - Originally used Cloudinary for photo storage but in the process of moving to S3.
 
-add photo upload capability for groups
+Implementing unit tests
 
-be able to select multiple musicians for one photo
+Improving Design
 
-In the future-- make timelines of players for each instrument in the group
+# Future features:
+
+Photo upload capability for groups
+
+Ability to select multiple musicians for one photo
+
+Timelines of players for each instrument in the group
