+++ package.json
@@ -6,7 +6,7 @@
   },
   "private": true,
   "scripts": {
-    "start": "node ./bin/www",
+    "start": "nodemon ./bin/www",
     "devstart": "nodemon ./bin/www",
     "test": "mocha"
   },
