+++ routes/index.js
@@ -4,19 +4,23 @@ var passport = require('passport')
 var Account = require('../models/account')
 var fs = require('fs')
 
-//handle photo upload
-var testS3 = require('../controllers/testupload')
+//handle photo upload with S3
+var testS3 = require('../test/testupload')
+var S3 = require('../S3')
+
+//handle file upload
 var multer = require('multer')
 var storage = multer.memoryStorage()
 var upload = multer({ storage: storage })
 var Image = require('../models/image')
-var uploadToS3 = require('../controllers/testupload')
-var cloudinary = require('cloudinary')
-cloudinary.config({
-  cloud_name: process.env.CLOUDINARY_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET
-})
+
+//storage has moved from cloudinary to S3
+// var cloudinary = require('cloudinary')
+// cloudinary.config({
+//   cloud_name: process.env.CLOUDINARY_NAME,
+//   api_key: process.env.CLOUDINARY_API_KEY,
+//   api_secret: process.env.CLOUDINARY_API_SECRET
+// })
 
 var musician_controller = require('../controllers/musicianController')
 var Musician = require('../models/musician')
@@ -64,10 +68,12 @@ router.get('/logout', function(req, res) {
     req.logout()
     res.redirect('/')
 })
-//Handle photo uploads
+//Test S3 connection
 router.get('/upload/test', function(req,res,next) {
   testS3.test()
 })
+
+//Photo upload form
 router.get('/upload', function(req, res, next) {
   Musician.find()
   .sort([['name', 'ascending']])
@@ -77,6 +83,8 @@ router.get('/upload', function(req, res, next) {
   })
 })
 router.post('/upload', upload.single('photoToUpload'), function (req, res, next) {
+  //I've moved the photo database from cloudinary to S3
+
   // cloudinary.uploader.upload(req.file.path, function(result){
   //   var image = new Image ({
   //     musician: req.body.musician,
@@ -88,28 +96,28 @@ router.post('/upload', upload.single('photoToUpload'), function (req, res, next)
   //     res.redirect('/catalog/musician/' + req.body.musician)
   //   })
   // })
+
   req.checkBody('musician', 'You must choose a musician').notEmpty()
   req.sanitize('caption').escape()
   req.sanitize('musician').escape()
   req.sanitize('caption').trim()
   var errors = req.validationErrors()
-  if (errors) {
+  if (errors) { //Rerender form with errors
     Musician.find()
     .sort([['name', 'ascending']])
-    .exec(function(err, list_musicians){
+    .exec(function(err, musicians){
       if (err) {return next(err) }
-      res.render('upload', {errors: errors, musicians: list_musicians, user: req.user})
-      return
+      res.render('upload', {errors: errors, musicians: musicians, user: req.user})
     })
   } else {
-    uploadToS3.upload(req.file, req.body, function() {
+    S3.upload(req.file, req.body, function() {
       var image = new Image ({
           musician: req.body.musician,
           caption: req.body.caption,
           key: req.file.originalname
       })
       image.save(function(err) {
-        if (err) {return err}
+        if (err) throw new Error(err)
         else res.redirect('/catalog/musician/' + req.body.musician)
       })
     })
