+++ test/test.user.js
@@ -1,45 +1,46 @@
-var should = require("should") 
-var mongoose = require('mongoose') 
-var Account = require("../models/account.js") 
-var db 
+var should = require("should")
+var mongoose = require('mongoose')
+var Account = require("../models/account.js")
+var db
 
 describe('Account', function() {
 
     before(function(done) {
-        db = mongoose.connect('mongodb://localhost/test') 
-            done() 
-    }) 
+        db = mongoose.connect('mongodb://localhost/test')
+            done()
+    })
 
     after(function(done) {
-        mongoose.connection.close() 
-        done() 
-    }) 
+        mongoose.connection.close()
+        done()
+    })
 
     beforeEach(function(done) {
         var account = new Account({
             username: '12345',
             password: 'testy'
-        }) 
+        })
 
         account.save(function(error) {
-            if (error) 
-            else 
-            done() 
-        }) 
-    }) 
+            if (error) throw new Error(error)
+            else
+            done()
+        })
+    })
 
     it('find a user by username', function(done) {
         Account.findOne({ username: '12345' }, function(err, account) {
-            account.username.should.eql('12345') 
-            
-            done() 
-        }) 
-    }) 
+            if (err) throw new Error(err)
+            account.username.should.eql('12345')
+
+            done()
+        })
+    })
 
     afterEach(function(done) {
         Account.remove({}, function() {
-            done() 
-        }) 
-     }) 
+            done()
+        })
+     })
 
-}) 
+})
